Runtime changed to load module's code, and rewrite them by embedding bid slots.

+++ runtime/src/io/stat.c
@@ -3,7 +3,7 @@
 #include <errno.h>
 
 int stat(const char * restrict path, struct stat * restrict buf) {
-  int rc = (int)__syscall2(SYS_stat, path, buf);
+  int rc = (int)__syscall2(SYS_stat, (long)path, (long)buf);
   if (rc < 0) {
     errn = -rc;
     rc = -1;
@@ -12,7 +12,7 @@ int stat(const char * restrict path, struct stat * restrict buf) {
 }
 
 int fstat(int fd, struct stat * buf) {
-  int rc = (int)__syscall2(SYS_fstat, fd, buf);
+  int rc = (int)__syscall2(SYS_fstat, fd, (long)buf);
   if (rc < 0) {
     errn = -rc;
     rc = -1;
